Move ballerina type extension to separate package
